mpegmap/HTMLSourcingInbandTracks-3.1-1.html

<!doctype html>
<html>
  <head>
    <title>TC 5.1.4-1 TextTrackList Order</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS - UA creates ordered TextTrackList.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Script Version: BETA -->
    <!-- About the Media Resource

         To verify DOM elements the details of the test media must be known prior to test execution.
         This test script is bound to the test media file sourced in the script.
         Sourcing a differnt test media file, or modifying that file, may invalidate the test.

         The test media is an MPEG-2, Single Program Transport Stream
         containing two EISS streams, and one program insertion cue stream.
         In test media file the PMT 'ES apperance order'
         is the same as the assending order of ES PIDs.
    -->
  </head>
  <body>
    <p>
      <a href="http://www.w3.org/TR/html5/references.html#refsINBANDTRACKS">Spec
      Reference [INBANDTRACKS]</a>
    </p>

    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>
    <script>
      var test = async_test("TextTrackList order.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "UserPrivateStreams.ts";

      // PMT ES appearance order
      var pids = ["sn-1", "1768", "1770"];
 
      vid.addEventListener("playing", function() {
        vid.pause();

        test.step(function() {
          // Make sure the tracks are in order
          var next = 0;
          for (var i = 0; i < vid.textTracks.length; i++) {
            if (pids.indexOf(vid.textTracks[i].id) !== -1) {
              assert_equals(pids[next], vid.textTracks[i].id, "TextTrackList Order");
              next++;
            }
          }
        }, "TextTrackListOrder.");

        test.done();
      }, false);
    </script>
  </body>
</html>
